requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
constantly==15.1.0
crayons==0.4.0
cryptography==3.3.1
cv==1.0.0
Cython==0.29.21
defusedxml==0.6.0
discord.py==1.5.1
distlib==0.3.1
Django==3.1.3
django-allauth==0.44.0
django-appconf==1.0.4
django-decouple==2.1
django-etc==1.2.0
django-grappelli==2.14.2
django-hitcount==1.3.3
django-tagging-autocomplete==0.5.1
django-taggit==1.3.0
django-taggit-autosuggest==0.3.8
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
filelock==3.0.12
GPUtil==1.4.0
h5py==3.1.0
idna==2.10
incremental==17.5.0
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
Keras==2.4.3
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mpmath==1.1.0
multidict==4.7.6
mypy-extensions==0.4.3
numpy==1.19.3
oauthlib==3.1.0
pathspec==0.8.1
pilkit==2.0
pillow==8.0.1
psutil==5.7.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
pylint==2.5.3
pymongo==3.11.2
python-decouple==3.3
python3-openid==3.2.0
pytorch3d==0.0.1
pytz==2020.4
pywin32==300
PyYAML==5.3.1
pyzmq==20.0.0
razorpay==1.2.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.5.4
selenium==3.141.0
Simon==0.7.0
simons==0.0.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.1.10
sqlparse==0.4.1
sympy==1.8.dev0
tabulate==0.8.7
tagging==0.2.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
webdriver-manager==3.2.2
whatsapp-web==0.0.1
WMI==1.5.1
wrapt==1.12.1
yarg==0.1.9
yarl==1.5.1
zope.interface==5.2.0
